+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## v0.11.12
+- Fix prune files with query params url.
 - Fix issue when get entity parent owner in sync actions with inserts.
 
 ## v0.11.11